fix: avoid unlink inode when getting a net error
When creating a file, the client will create inode first and then send a
create dentry request. If create dentry request fails due to a net
error (e.g. send the request to metanode but timed out when getting
response), there is a tiny chance that the dentry is created
successfully on metanode. Thus we cannot unlink the inode created
previously because we would leave the orphan inode rather than creating
an orphan dentry.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>